third_party/blink/web_tests/platform/mac/external/wpt/css/css-conditional/container-queries/scroll-state/scroll-state-snapped-snap-changing-expected.txt

This is a testharness.js-based test.
[FAIL] scroll-state(snapped) matching #snapped2 while scrolling
  assert_equals: snapchanging has been called for #snapped2 expected (object) Element node <div id="snapped2" class="snapped">\n  <span id="target2">... but got (undefined) undefined
Harness: the test ran to completion.
